+++ src/main/java/ml/duncte123/skybot/objects/pairs/LongLongPair.java
@@ -1,0 +1,37 @@
+/*
+ * Skybot, a multipurpose discord bot
+ *      Copyright (C) 2017 - 2020  Duncan "duncte123" Sterken & Ramid "ramidzkh" Khan & Maurice R S "Sanduhr32"
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published
+ * by the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package ml.duncte123.skybot.objects.pairs;
+
+public class LongLongPair {
+    private final long first;
+    private final long second;
+
+    public LongLongPair(long left, long right) {
+        this.first = left;
+        this.second = right;
+    }
+
+    public long getFirst() {
+        return first;
+    }
+
+    public long getSecond() {
+        return second;
+    }
+}
